requirements.txt
astroid==2.3.3
-e git+https://github.com/shakthydoss/abc.git@7aa0defbbd3237812d9d831da6b5c375279d0947#egg=genie&subdirectory=tools
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
pdfkit==0.6.1
Pillow==7.0.0
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
